Add tests for coordinate transformations
The currently set tolerance for deviations between expected and result
coordinates of 0.01 m seem too ambitious. 19 of 25 tests fail with
differences between 0.03 and 4.09 m. It is not obvious whether this is
due to inaccurate transformation during generation of expected results
with PostGIS and/or in our implementation respectively the underlying
GeoTools or even calculation of difference within the test implementation
itself.

Test coverage for transforming from/to coordinates expressed in
latitudinal/longitudinal degrees, minutes and seconds is still missing.