Use subquery filtering for export of selected entries
This allows to omit maintaining a map of query column aliases to
table columns, which also seems to be a bit too much to assume this
as stable because queries and their column aliases can be freely
"configured" in the database. Although we have some semantics
associated to specific ID-column names (see README.markdown) to
enable the client to make decisions about functionality for specific
data types, there's no obligation to hard-wire this also in server
code. Instead, introduce a new filter type that is rendered into a
WHERE-clause to filter a subquery based on its column aliases.