Fix formatting of arc minutes and degrees
Previously, the format string for seconds missed that 'width' (which was
given as 2) in a format string is, as per documentation of
java.util.Formatter, the "minimum number of characters to be written to
the output". Since it was less than the given precision, the intended
zero padding of single-digit values never took place.

While at it, try to make the code a bit more comprehensible.