Simplify class hierarchy for database access
The implementation provided different classes for read-only and write
access for use in service implementations as a measure of protection in
services that do not require write access. This became obsolete with more
fine-grained authorization through Authorizer implementations. Therefore,
boil database access implementation down to a single class "Repository"
providing the whole functionality.