Remove unused filter methods
The signature with JsonArray is unused since rev. ccf7cdef, so there
might be a problem with it, though nothing is documented in the respective
commit message. The signatures with size and start parameters seem to
have been intended for paging but are unused as well. If paging is ever
needed at this level, it should be pushed down to the database, anyhow,
to avoid fetching the whole data set each time the method is called.