Reactivate tests for MesswertService
This is mostly cleanup of foreign key values and updating attribute names.
The attribute "messwert" is removed from dbUnit_messwert.json because it
is well possible to seed the database with it, but the created object will
not pass validation in the HasMesswert validation rule during the update
operation: one of "messwert" and "messwert_nwg" has to be null. The
database schema and validation rule do not match here.