Fix comparison of integer values
To compare the int values represented by Integer objects, the equals() method
has to be used. Also reported as RC_REF_COMPARISON in the SpotBugs report.
Previously, the respective tests in StammdatenTest failed.
While at it, simplify the logic to check for duplicates that was added in
rev. ebebadbc and 8f1c4169.