Add basic constraints to tag tables
In passing, make the schema notation more compact by avoiding explicit
mentioning of PostgreSQL defaults.

Using boolean instead of Boolean in the model class is essentially
expressing 'NOT NULL DEFAULT false' in Java, so we do not need to use
something like org.hibernate.annotations.DynamicInsert for the DEFAULT
to take effect. The removed (and albeit unnecessary) extra roundtrip
to the database in TagService is no longer possible, because a
'tagzuordnung' without tag is not allowed anymore (and albeit makes
no sense).