Avoid Internal Server Error on creation of user selection
Since rev. 356573dc2, stamm.grid_column_values.filter_is_null has a
NOT NULL constraint. Since the client does not send the respective field
to ColumnValueService on creation of a new user-specified selection,
this led to an Internal Server Error due to the field being null. Thus
change the fields type to boolean, which matches the database
DEFAULT false. Do the same for other affected fields in GridColumnValue,
although they are send by the client in the above mentioned case.
This also introduces simplification to some conditionals in QueryTools
and makes them less ambiguous.