Fix updates in transactional context
The ORM does neither immediately issue an UPDATE in the database during
EntityManager.merge() nor does it refresh the to-be-updated object.
The work-around for this has been to use an additional Repository.getById()
to fetch the updated entry from the database in many places in the code.
Since rev. 8c0ffdef5, this does no longer work because the additional getById()
fetches the entry from the same persistence context, still being out-of-sync
with the database. Thus, flush and refresh explicitly and remove the former
work-around in all places I could find it.
A symptom of the entries being out-of-sync was that PUT requests received
data with outdated modification timestamps, which led to errors in following
PUT requests for the same resource. Therefore, test this in ServiceTest.update().