Count references to Messprogramm objects directly in the model
Makes this property really related to data. Before, if a client issued a
POST request with a wrong value (e.g. other than zero, because a new object
cannot be referenced already), that value was kept in the response
if the response still referenced the very same instance created from the data
in the POST request.
In passing, fix some code style issues.