Simplify export file creation
Using File.createTempFile() instead of a hard-coded absolute base path
should make the code more portable. Not using the ID of the export in
the filename anymore is a first step in making job creation simpler using
injection rather than constructor, which in turn is part of moving the
concurrent processing towards using Java EE concurrency utils to enhance
thread-safety.