Fix rev. e1684432da0a3944fcc48012cc01d7ab47da17cf
Schema update scripts have to update the schema to the state that is defined
in the revision based on the previous state. Thus, tables that are introduced
with the revision have to be CREATEd, not ALTERed. Re-enable updates of
existing databases by removing the respective table modifications from schema
update #0026 and add the missing tables in a new schema update.