Implement authorization only once in Authorizer
This avoids having multiple implementations in service methods and
centralizes some basic input validation. Also remove weird query
construct in delete method: if the request is to delete an instance
specified by ID, delete that one and not the one that is accidentally
the first one returned by some query.