Fix deleting TagZuordnung
We need to fetch the entity with its ID before removing it to
avoid an INSERT caused by calling EntityManager.merge() in
Repository.delete(). In order to be able to use tagId in criteria
query, this field cannot be transient. Thus, simplify data model
at this point and in passing remove unnecessary method "toJson".