Restrict GET for Messung objects to those associated to given probeId
Getting all Messung objects can likely lead to a performance nightmare
(long running database query, memory shortage) if many entries are in
the database. Since there's currently no use-case for paging when
querying with a probeId, remove that functionality completely. The
check with Response.getSuccess() is obsolete, because it always
returns true.