Improve handling of empty strings as URL parameters
In case an URL parameter is bound to a type other than String and a
client sends a request with a parameter value being an empty string,
the server answers with HTTP status code 404. In case the parameter
is bound to String, an empty string is, by default, a valid value.
Thus, add validation where the value is actually expected to be
not empty.