Use generics instead of qualifiers to identify validators
Having the type of the object to be validated at hand will allow
to integrate standard Jakarta Bean Validation.
Since passing an object of wrong type to the validate method is
a programming error, in such cases no longer return a validation
error to the client. The occurring ClassCastException will be
more helpful to track down the cause.