Reduce code duplication
Build map of IDs with records at one place.

Note that the parameter columnsToInclude of Exporter.export()
relies on QueryExportJob.columnsToExport, which is never null.
Thus the null-fallback can be skipped and the type of parameter
queryResult can be released to Iterable, which allows passing
the collection view on the ID-records map.