Remove general purpose response wrapper
Since success/failure is generally handled at HTTP level, it is no
longer necessary to wrap responses in an object with a Boolean and
a String value indicating success. Were a response can contain a list
of successful as well as failed states, this is modeled using specific
return types. The total count of returned entities is only used in a
few places and is now as well modeled using specific return types.