Make tag type implicit
Tags are identifiable by name and either networkId, measFacilId
or none of them, depending on the type. Thus, the type is defined
by the occurrence of these keys and the redundant explicit type
was a source for contradiction.
Since tags are identifiable as described above, no extra UNIQUE
index for generated tags is necessary.