Fix validation of environment descriptor vs. environmental medium
For REI and §161 samples, a validation notification should be emitted in
case the specific matching rule returns an ambiguous result, but
nevertheless, a warning should be emitted if no EnvDescripEnvMediumMp
references the same EnvMedium as the given sample or measuring program.
The latter is now ensured instead of emitting a notification in this
case as well.