Prevent false warnings after import of sample with dependent objects
Fixes a regression of rev. 07a2acfc that caused warnings generated in a
first validation run of the sample, before import of dependent objects,
to be propagated to clients.
Changes to test data are necessary to be able to import location data
at the one hand, and to prevent other warnings at the other hand.