Restrict return value validation to warnings and notifications
Especially when validating large lists, this can be a huge performance
gain and avoid timeouts.
Also drops the assumption, that constraints in groups Warnings and
Notifications cannot be expected to work with Beans that also contain
errors. Since all constraints are implemented as Bean Validation
constraints, none of them should make assumptions about the group in
which they are used.