Drop status field from Measm
The field mapped a foreign key in the database that points to the most
current entry of the 1:n relation with StatusProt. This field is managed
by database triggers and should not otherwise be manipulated. This was
nevertheless possible even for a client by simply updating it via a
PUT request. Thus, drop the field and use the OneToOne-related StatusProt
instead, which represents the same entry, but without being susceptible
to manipulation.
Also drops an unnecessary null check in BaseAuthorizer, which did not
take into account that the database ensures that every persisted Measm
object has a related StatusProt entry.