Improve type-safety in authorization
Make Authorizer a parameterized type and ensure that authorized data objects
are of the type matching the type parameter of the Authorizer implementation.
This should prevent ClassCastExceptions due to arbitrary data objects being
passed to a not matching Authorizer implementation.
Also includes the UserInfo reference as a field in Authorizers in order
to avoid having to pass it in every method call.
Authorizer implementations are no longer public, because from outside the
package, they should be used through the Authorization interface.
The reason in case authorization is not successful is now transported as an
exception message. This is just to be able to implement the choice of
Authorizer implementation only once in HeaderAuthorization.doAuthorize()
respectively to minimize the code repeated for each type therein.