Set sample on deserialization of associated entities
Removes ambiguity between the field containing the associated sample
and the field containing the foreign key referencing the sample.
Before, a deserialized entity not yet attached to the persistence
context did have the foreign key field set, but not the association.
This led to various places where the associated sample had to be obtained
by explicitly loading it from the persistence context.