Rename refId and id to jobId.
In order to improve the documentation via OpenAPI path variables were
renamed from id to jobId to make the semantics more clear that a given
Id is referring to a job. Additionally the AsyncJobResponse uses jobId
instead of refId to make things more clear.

The renaming was reflected in the tests as well.