Introduce bidirectional association between sample and measurements
Use byte-code enhancement to release from (part of) the burden to keep
in sync both sides of the association. This requires the mappedBy
field to be visible to the Entity-class to which the mapping points.
Therefore, the access modifier is changed to protected in the superclass.

Since this is in preparation for a new JSON-based import format that will
allow importing samples including associated objects, introduce a provisional
service for uploading such samples. To preserve the behavior of the existing
services, make them ignore associated measurements and restrict the set of
cascaded operations.

While at it, cleanup JavaDoc a bit.