Make setting Site.plausibleReferenceCount thread-safe
Listeners should not save state as instance variables (see
https://docs.jboss.org/hibernate/orm/6.6/userguide/html_single/Hibernate_User_Guide.html#events-events).
Doing so with referencing an EntityManager via Repository in SiteListener
did not make any trouble so far. But it will do so as soon as the listener
will be used e.g. inside and outside of the transaction of service methods.

Using @Formula to retrieve a field value hampers using native SQL queries
to create entity instances, because said field is missing in the query
result. Therefore, the query in SiteService has been migrated to JPQL.

While at it, remove some unused fields, getters and setters from Site.