More helpful error message for BaseTest::assertResponseStatus.
The intention of the original code was that a response object
should be parsed and returned in case everything is okay
or to put out an assertion error message with details about the
error occured.

But in case something went wrong, i.e. the statuscode which was expected
was not the one returned the output of the assert statement could be
quirky (e.g. validationtests show unhelpful error messages).

This is an attempt to improve the situation.

Because the response object could only be read/parsed once the code is a bit clunky.
Details see: https://jakarta.ee/specifications/platform/10/apidocs/jakarta/ws/rs/core/response
»IllegalStateException - if the entity was previously fully consumed as an input stream,
 or if the response has been closed.«

Because the assertion statement needs the textual representation of the
response body the statement is wrapped up in an if statement assuring
this condition. Only then the response object is parsed to a string.
Given that error condition the code flow ends here with an Exception
and the return statement at the end of the method isn't executed and
doesn't parse the object for a second time.

In case everything works as expected the string generation for the
assertion could be skipped anyways.

So the original behaviour is preserved.

A more clean approach would have been to parse the response body
to an intermediary string and in case everything is okay leverage
an ObjectMapper to unmarshall it to an object again.

But

a) would this unmarshalling be unnecessary for a great number of
cases. The assertion is only for defensive purposes and
error cases shouldn't be common.

b) Is the whole ObjectMapper interface incompatible to the GenericType<T>
type such that the function signature would have changed with the
according ripples throughout the test code. And to refactor that for a corner case
seems to be too much effort.

This solution here I believe is the most pragmatic solution
done with the least effort.