Site: map and img columns are now declared @Column(updatable=false).
Previously there was an error due to a side effect of the columns being
@JsonbTransient. When the site object was updated (e.g. via
SiteService) the value of the img or map column of a Site-object was set to null.
The reason for that is: they were filtered out via @JsonbTransient and so existing
data was inadvertedly deleted.

In order to preserve existing data during ordinary repository calls the columns are now defined as
non-updatable.

Regardless there is the possibility of JPQL-Queries to update this specific columns.
This was used to (re-) set the according data.