Added new docker build target to developer documentation.
The way of how to build the client for production and development
changed. In order to build the correct image one has to specify the
build target. Without any specification the build defaults to
production. So we have to add the target option "development" in order
to get a development build.