Fix authorizing modification of managed entities
The check, whether the unmodified state was allowed to be modified,
was actually made based on the modified state. Firstly, because of
Hibernate issuing a flush before querying the state from the database.
This is now prevented by setting the flushing strategy to manual.
Secondly, because querying the state of the modified managed entity
does issue an SQL query, but nevertheless returns the modified
instance. This is now prevented by detaching the modified entity
prior to the query.