Remove unused fields
lastMod is unused in server code and there are no known clients
that use this field in case of the modified classes. If it turns
out there are any clients that need it, it can be re-introduced
at a central place as truly read-only (insertable=false,
updatable=false, no setter), since it is and should be under
control of the database.