Fix determining last status protocol entry
Add a sequence number per measurement to ensure correct ordering even
if multiple entries have the same timestamp. The latter can happen in
case the entries have been created with the same import. Breaking ties
using statusVal did not work as intended.
Fetching the latest status mapping ID with a named query also has the
advantage, that it does not depend on loading all protocol entries of
a measurement.
Ordering Measm.statusProts makes it easy for clients to identify the
latest entry without client-side sorting as a replacement for the
dropped Measm.statusProt.
The fixed ordering also fixes the read-only-flag at imported samples,
which is no longer false, even if imported status implies true.
Nevertheless, ignore in import tests, because the flag is of no
importance for import.